Update load-sea-constants-modules.R
For FRAC_3 module, one of the features of the problems was named (and therefore coded) incorrectly, though the logic was correct. Problems can be congruent or incongruent. A problem is congruent when the larger fraction also has the larger components (e.g., 3/7 > 2/6). A problem is incongruent when the larger fraction has smaller components (e.g., 3/4 > 4/7). In the old code, this was incorrectly named num_larger.